// src/lib/generated/nexusTypes.gen.ts
export interface NexusGenInputs {
  userSignupPayloadType: { // input type
    email?: string | null; // String
    name?: string | null; // String
    password?: string | null; // String
  }
}

export interface NexusGenEnums {
}

export interface NexusGenScalars {
  String: string
  Int: number
  Float: number
  Boolean: boolean
  ID: string
}

export interface NexusGenObjects {
  Mutation: {};
  Query: {};
  userList: { // root type
    email?: string | null; // String
    id?: string | null; // String
    name?: string | null; // String
  }
  userSignupPayloadSuccessType: { // root type
    email?: string | null; // String
    id?: string | null; // String
    message?: string | null; // String
  }
}

export interface NexusGenInterfaces {
}

export interface NexusGenUnions {
}

export type NexusGenRootTypes = NexusGenObjects
// ...
export interface NexusGenArgTypes {
  Mutation: {
    signup: { // args
      data?: NexusGenInputs['userSignupPayloadType'] | null; // userSignupPayloadType
    }
  }
}